requirements.txt
pandas
scikit-learn
spacy
pytesseract
Pillow
pdf2image
reportlab
